requirements.txt
carla-birdeye-view
gputil
psutil